Added support to better work with "shiny"
Added methods `get_bounds`, `set_bounds`, and `get_frame` to
Handle getting/setting `self.bounds_data` or `self.frame`.  These methods
include the ability to work without multiindexes.  In the 'get' methods, there
is a option `reset_index=False/True` to get a version of the frame with/without
multiindexes. In the `set_bounds` there is an option `set_index` which will set
the multiindex for `bounds_data`.  Note that the default for `set_index` is to
try to figure out if the index should be set or not.